config db

+++ server/prisma/migrations/20230325015344_create_database/migration.sql
@@ -2,11 +2,11 @@
 CREATE TABLE "CurrentWeather" (
     "id" TEXT NOT NULL PRIMARY KEY,
     "city" TEXT NOT NULL,
-    "lat" INTEGER NOT NULL,
-    "lon" INTEGER NOT NULL,
+    "lat" REAL NOT NULL,
+    "lon" REAL NOT NULL,
     "dt" INTEGER NOT NULL,
-    "temp" INTEGER NOT NULL,
-    "feels" INTEGER NOT NULL,
+    "temp" REAL NOT NULL,
+    "feels" REAL NOT NULL,
     "sunrise" INTEGER NOT NULL,
     "sunset" INTEGER NOT NULL
 );
